@@ -67,7 +67,7 @@ class BannedAdoptersController < ApplicationController
 
   def update
     @banned_adopter = BannedAdopter.find_by(id: params[:id])
-    if @banned_adopter.update_attributes(banned_adopter_params)
+    if @banned_adopter.update(banned_adopter_params)
       flash[:success] = 'Record updated.'
       redirect_to banned_adopters_path
     else
