@@ -142,7 +142,7 @@ class User < ApplicationRecord
 
   has_many :assignments, class_name: 'Adopter', foreign_key: 'assigned_to_user_id'
   has_many :active_applications, -> { where(status: ['new', 'pend response', 'workup', 'ready for final', 'approved']) }, class_name: 'Adopter', foreign_key: 'assigned_to_user_id'
-  belongs_to :mentor, class_name: 'User', foreign_key: 'mentor_id'
+  belongs_to :mentor, class_name: 'User', foreign_key: 'mentor_id', optional: true
   has_many :mentees, class_name: 'User', foreign_key: 'mentor_id'
 
   before_validation :sanitize_postal_code
