@@ -3,12 +3,13 @@
 # Version of your assets, change this if you want to expire all your assets.
 Rails.application.config.assets.version = '1.0'
 
-# Add additional assets to the asset load path
+# Add additional assets to the asset load path.
 # Rails.application.config.assets.paths << Emoji.images_path
 # include the sponsors directory here so it's not necessary to prepend "sponsors/" where the image is used
 Rails.application.config.assets.paths << Rails.root.join("app", "assets", "images", "sponsors")
 
 # Precompile additional assets.
-# application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
-# Rails.application.config.assets.precompile += %w( search.js )
+# application.js, application.css, and all non-JS/CSS in the app/assets
+# folder are already added.
+# Rails.application.config.assets.precompile += %w( admin.js admin.css )
 Rails.application.config.assets.precompile += %w( application_pdf.css application_bs41.css application_bs41.js galleria-1.5.7.classic.js galleria-1.5.7.classic.css)
